Small Code refactor
- Moved BLE Manager into it's own class
- Moved ReceivedPacket into it's own class and files
- Made the function style more consistent across them all
