+++ Software/Human-Zombie-Device/Ble_Manager.h
@@ -1,26 +1,15 @@
 #ifndef BLE_MANAGER_H
 #define BLE_MANAGER_H
 
-void init_ble();
-void scan_for_ztags();
-void print_all_ztags_found();
-
-
-// Data structure for storing found data
-class ReceivedPacket
+class Ble_Manager
 {
 private:
-    int rssi;
-    String message;
+    /* data */
+    void processZtags();
 public:
-    
-    int getRssi();
-    String getMessage();
-
-    void setRssi(int rssi);
-    void setMessage(String message);
+    void initBle();
+    void scanForZtags();
+    void pritnAllZTagsFound();
 };
-
-
 
 #endif   // BLE_MANAGER_H
